+++ NEWS.md
@@ -19,8 +19,10 @@ NEW FEATURES (Antares v8.6) :
 
 NEW FEATURES :
 
+* Add deduplicateScenarioBuilder() function to keep the last value if a key is duplicated in settings/scenariobuilder.dat
 * Add writeIniHydro() function to make easier the edition of the input/hydro/hydro.ini file
 * Call writeIniHydro() in createArea() and removeArea()
+
 
 ### Breaking changes